pyproject.toml
[project]
name = "project-sk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.0.0",
    "google-search-results>=2.4.2",
    "httpx>=0.27.0",
    "langchain>=0.2.0",
    "langchain-anthropic>=0.1.0",
    "langchain-community>=0.2.0",
    "langchain-core>=0.2.0",
    "langchain-google-genai>=1.0.0",
    "langchain-groq>=0.1.0",
    "langchain-huggingface>=0.0.3",
    "langchain-ollama>=0.1.0",
    "langchain-openai>=0.1.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
    "python-telegram-bot>=20.0",
    "requests>=2.31.0",
    "selenium>=4.18.0",
    "wikipedia>=1.4.0",
]
